fix(agentless-scanning): correct caller_arn for assumed roles (#75)
When a role is assumed, the caller identity ARN is an ephemeral
session-specific role. Rather than this role in IAM policies, we should
use the underlying role that was assumed. This role can be retrieved
with `aws_iam_session_context`. More info can be found in [this][1] AWS
provider issue.

[1]: https://github.com/hashicorp/terraform-provider-aws/issues/28381